+++ src/app/CookiesWrapper.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { CookiesProvider } from 'react-cookie';
+
+export function CookiesWrapper({ children }) {
+  return <CookiesProvider>{children}</CookiesProvider>;
+}