+++ src/app/layout.tsx
@@ -13,6 +13,9 @@ import { AppFooter } from '@/components/AppFooter';
 import { ServiceWorkerLoader } from '@/components/ServiceWorkerLoader';
 import dynamic from 'next/dynamic';
 import { cookies } from 'next/headers';
+import { AnalyticsScript } from '@/components/AnalyticsScript';
+import { CookiesProvider } from 'react-cookie';
+import { CookiesWrapper } from './CookiesWrapper';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -33,6 +36,8 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const currentCookies = await cookies();
+  console.log(currentCookies.get('wagerize-analytics-enabled'));
   return (
     <html lang='en'>
       <head>
@@ -40,16 +45,21 @@ export default async function RootLayout({
           rel='manifest'
           href='manifest.json'
         />
+        <AnalyticsScript
+          enabled={currentCookies.get('wagerize-analytics-enabled').value === 'true'}
+        />
       </head>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col w-full h-screen flex-1`}>
-        <AuthProvider>
-          <UserProvider>
-            <AppHeader />
-            {children}
-            <AppFooter />
-          </UserProvider>
-        </AuthProvider>
+        <CookiesWrapper>
+          <AuthProvider>
+            <UserProvider>
+              <AppHeader />
+              {children}
+              <AppFooter />
+            </UserProvider>
+          </AuthProvider>
+        </CookiesWrapper>
         <Toaster position='top-center' />
       </body>
     </html>