+++ src/components/util/PassProps.tsx
@@ -5,6 +5,10 @@ export function PassProps({ children, ...props }) {
     React.cloneElement(child, {
       ...child.props,
       ...props,
+      onClick: e => {
+        child.props.onClick && child.props.onClick(e);
+        props.onClick && props.onClick(e);
+      },
       className: [child.props.className, props.className].join(' ').trim(),
     })
   );